fix: Use `UpdateButton` on SnapPage
