+++ packages/app_center/lib/manage/manage_snap_tile.dart
@@ -1,6 +1,7 @@
 import 'package:app_center/l10n.dart';
 import 'package:app_center/layout.dart';
 import 'package:app_center/manage/manage_l10n.dart';
+import 'package:app_center/manage/update_button.dart';
 import 'package:app_center/snapd/snap_action.dart';
 import 'package:app_center/snapd/snapd.dart';
 import 'package:app_center/store/store.dart';
@@ -260,7 +261,7 @@ class _ButtonBar extends ConsumerWidget {
         )
       else ...[
         if (showUpdateButton)
-          _UpdateButton(snapModel: snapModel, activeChangeId: activeChangeId),
+          UpdateButton(snapModel: snapModel, activeChangeId: activeChangeId),
         if (!showUpdateButton && canOpen)
           OutlinedButton(
             onPressed: snapLauncher.open,
@@ -273,71 +274,5 @@ class _ButtonBar extends ConsumerWidget {
           ),
       ],
     ];
-  }
-}
-
-class _UpdateButton extends ConsumerWidget {
-  const _UpdateButton({
-    required this.snapModel,
-    required this.activeChangeId,
-  });
-
-  final AsyncValue<SnapData> snapModel;
-  final String? activeChangeId;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final l10n = AppLocalizations.of(context);
-    final shouldQuitToUpdate =
-        snapModel.valueOrNull?.localSnap?.refreshInhibit != null;
-    final snap =
-        snapModel.valueOrNull?.localSnap ?? snapModel.valueOrNull?.storeSnap;
-
-    if (shouldQuitToUpdate) {
-      return const QuitToUpdateNotice();
-    } else {
-      return OutlinedButton(
-        onPressed: activeChangeId != null || !snapModel.hasValue
-            ? null
-            : ref.read(snapModelProvider(snap!.name).notifier).refresh,
-        child: Row(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            const Icon(YaruIcons.download),
-            const SizedBox(width: kSpacingSmall),
-            Text(
-              l10n.snapActionUpdateLabel,
-              maxLines: 1,
-              overflow: TextOverflow.ellipsis,
-            ),
-          ],
-        ),
-      );
-    }
-  }
-}
-
-@visibleForTesting
-class QuitToUpdateNotice extends StatelessWidget {
-  const QuitToUpdateNotice({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = AppLocalizations.of(context);
-    final colorScheme = Theme.of(context).colorScheme;
-
-    return Row(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        Icon(YaruIcons.warning_filled, color: colorScheme.warning),
-        const SizedBox(width: 8),
-        Text(
-          l10n.managePageQuitToUpdate,
-          maxLines: 1,
-          overflow: TextOverflow.ellipsis,
-          style: Theme.of(context).textTheme.bodyMedium,
-        ),
-      ],
-    );
   }
 }
